Created three folders to separate the code, and create a class for app instead of a function.
