+++ src/app/App.css
@@ -1,3 +1,8 @@
+body {
+  background-color: black;
+}
+
+
 .App {
   text-align: center;
 }
